backend/requirements.txt
Django==5.1.4
psycopg==3.2.3
transformers==4.47.1
requests==2.32.3
youtube-transcript-api==0.6.3
yt-dlp==2023.12.23
pytesseract==0.3.13
pymupdf==1.25.1
beautifulsoup4==4.12.3
torch==2.5.1
djangorestframework==3.15.2
django-cors-headers==4.6.0